+++ packages/views/locales/en/runtimes.json
@@ -257,7 +257,6 @@
     "col_health": "Health",
     "col_owner": "Owner",
     "col_agents": "Agents",
-    "col_workload": "Workload",
     "col_cost": "Cost · 7d",
     "col_cli": "CLI",
     "cost_delta_flat": "flat",
