+++ packages/views/runtimes/components/runtime-list.tsx
@@ -1,8 +1,9 @@
 "use client";
 
-import { useMemo } from "react";
+import { useMemo, useState } from "react";
+import { Globe, MoreHorizontal, Trash2 } from "lucide-react";
+import { toast } from "sonner";
 import { useQuery } from "@tanstack/react-query";
-import { getCoreRowModel, useReactTable } from "@tanstack/react-table";
 import type {
   Agent,
   AgentRuntime,
@@ -15,12 +16,91 @@ import {
   agentListOptions,
   memberListOptions,
 } from "@multica/core/workspace/queries";
-import { agentTaskSnapshotOptions } from "@multica/core/agents";
-import { paths, useWorkspaceSlug } from "@multica/core/paths";
-import { DataTable } from "@multica/ui/components/ui/data-table";
-import { useNavigation } from "../../navigation";
-import { type RuntimeRow, createRuntimeColumns } from "./runtime-columns";
+import { agentTaskSnapshotOptions, deriveWorkload } from "@multica/core/agents";
+import {
+  deriveRuntimeHealth,
+  runtimeUsageOptions,
+} from "@multica/core/runtimes";
+import { useWorkspacePaths } from "@multica/core/paths";
+import { Button } from "@multica/ui/components/ui/button";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@multica/ui/components/ui/dropdown-menu";
+import {
+  ListGrid,
+  ListGridCell,
+  ListGridHeader,
+  ListGridHeaderCell,
+  ListGridRow,
+} from "@multica/ui/components/ui/list-grid";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+} from "@multica/ui/components/ui/tooltip";
+import { AppLink } from "../../navigation";
+import { ActorAvatar } from "../../common/actor-avatar";
+import { useViewingTimezone } from "../../common/use-viewing-timezone";
+import { workloadConfig } from "../../agents/presence";
+import { ProviderLogo } from "./provider-logo";
+import { HealthIcon, useHealthLabel } from "./shared";
+import { DeleteRuntimeDialog } from "./delete-runtime-dialog";
+import {
+  computeCostInWindow,
+  formatLastSeen,
+  isSelfHealingRuntime,
+  pctChange,
+} from "../utils";
+import { splitRuntimeName } from "./runtime-machines";
 import { useT } from "../../i18n";
+
+// The machine detail's runtimes table on the shared ListGrid. Paradigm
+// pieces are taken À LA CARTE here: subgrid template + var-width tracks +
+// two-zone responsiveness (the detail pane gets squeezed by the machine
+// list, so the container-driven core-set collapse matters more than on the
+// full-width pages), but NO virtualization / sorting / filters / column
+// toggles / batch selection — a machine hosts 1-5 runtimes, those would all
+// be dead weight, and batch-deleting runtimes (a cascade-confirm heavy
+// operation) is deliberately not offered.
+const GRID_COLS =
+  "grid-cols-[0.75rem_minmax(120px,1fr)_var(--rtc-health)_1.75rem_0.75rem] " +
+  "@2xl:grid-cols-[0.75rem_minmax(140px,1fr)_var(--rtc-health)_var(--rtc-owner)_var(--rtc-agents)_var(--rtc-cost)_var(--rtc-cli)_1.75rem_0.75rem]";
+
+const COLUMN_WIDTHS = {
+  // Health folds the workload in as a suffix ("Healthy · 2 running") —
+  // same merge as the agents list's status cell.
+  health: 176,
+  owner: 96,
+  agents: 92,
+  cost: 96,
+  cli: 112,
+} as const;
+
+// Fixed tracks (edges 12+12, name min 140, kebab 28) plus the 8 gap-x-3
+// gaps between the wide template's 9 tracks (zero-width tracks still
+// carry gaps).
+const FIXED_TRACKS_WIDTH = 192 + 8 * 12;
+
+function columnTrackVars(showOwner: boolean): React.CSSProperties {
+  const minWidth =
+    FIXED_TRACKS_WIDTH +
+    COLUMN_WIDTHS.health +
+    (showOwner ? COLUMN_WIDTHS.owner : 0) +
+    COLUMN_WIDTHS.agents +
+    COLUMN_WIDTHS.cost +
+    COLUMN_WIDTHS.cli;
+  return {
+    "--rtc-health": `${COLUMN_WIDTHS.health}px`,
+    "--rtc-owner": showOwner ? `${COLUMN_WIDTHS.owner}px` : "0px",
+    "--rtc-agents": `${COLUMN_WIDTHS.agents}px`,
+    "--rtc-cost": `${COLUMN_WIDTHS.cost}px`,
+    "--rtc-cli": `${COLUMN_WIDTHS.cli}px`,
+    "--rtc-minw": `${minWidth}px`,
+  } as React.CSSProperties;
+}
 
 interface RuntimeWorkload {
   agentIds: string[];
@@ -33,6 +113,13 @@ const EMPTY_WORKLOAD: RuntimeWorkload = {
   runningCount: 0,
   queuedCount: 0,
 };
+
+export interface RuntimeRow {
+  runtime: AgentRuntime;
+  ownerMember: MemberWithUser | null;
+  workload: RuntimeWorkload;
+  canDelete: boolean;
+}
 
 // Per-runtime workload snapshot — agent IDs serving this runtime (drives
 // the avatar stack; .length doubles as the agent count) plus task counts
@@ -69,6 +156,294 @@ export function buildWorkloadIndex(
   return result;
 }
 
+// ---------------------------------------------------------------------------
+// Cells
+// ---------------------------------------------------------------------------
+
+function RuntimeNameCell({ runtime }: { runtime: AgentRuntime }) {
+  const { base: baseName } = splitRuntimeName(runtime.name);
+  return (
+    <ListGridCell className="gap-2">
+      <div className="flex h-8 w-8 shrink-0 items-center justify-center">
+        <ProviderLogo provider={runtime.provider} className="h-5 w-5" />
+      </div>
+      <div className="flex min-w-0 flex-1 items-center gap-1.5">
+        <span className="block min-w-0 shrink truncate text-sm font-medium">
+          {baseName}
+        </span>
+        <VisibilityBadge runtime={runtime} />
+      </div>
+    </ListGridCell>
+  );
+}
+
+// Only public is worth a badge — private is the default and rendering a
+// `🔒 Private` chip on every row turns the whole column into noise.
+function VisibilityBadge({ runtime }: { runtime: AgentRuntime }) {
+  const { t } = useT("runtimes");
+  if (runtime.visibility !== "public") return null;
+  return (
+    <Tooltip>
+      <TooltipTrigger
+        render={
+          <span className="inline-flex shrink-0 items-center gap-0.5 rounded bg-info/10 px-1 text-[10px] font-medium text-info">
+            <Globe className="h-2.5 w-2.5" />
+            {t(($) => $.detail.visibility_label.public)}
+          </span>
+        }
+      />
+      <TooltipContent>
+        {t(($) => $.detail.visibility_hint.public)}
+      </TooltipContent>
+    </Tooltip>
+  );
+}
+
+// Health with the workload folded in as a suffix ("Healthy · 2 running").
+// Offline-ish rows skip the suffix — health already says it all. Online
+// idle rows also skip it: idle is the unremarkable default.
+function HealthCell({
+  runtime,
+  workload,
+  now,
+}: {
+  runtime: AgentRuntime;
+  workload: RuntimeWorkload;
+  now: number;
+}) {
+  const { t: tAgents } = useT("agents");
+  const labelOf = useHealthLabel();
+  const health = deriveRuntimeHealth(runtime, now);
+  const offline = health === "offline" || health === "about_to_gc";
+  const lastSeen = formatLastSeen(runtime.last_seen_at);
+
+  const wlState = deriveWorkload({
+    runningCount: workload.runningCount,
+    queuedCount: workload.queuedCount,
+  });
+  const wl = workloadConfig[wlState];
+  const counts =
+    wlState === "working"
+      ? workload.queuedCount > 0
+        ? `${workload.runningCount} +${workload.queuedCount}q`
+        : `${workload.runningCount}`
+      : wlState === "queued"
+        ? `${workload.queuedCount}`
+        : null;
+
+  return (
+    <ListGridCell className="gap-1.5">
+      <HealthIcon health={health} />
+      <span className="block min-w-0 truncate text-xs">
+        {labelOf(health)}
+        {health !== "online" && runtime.last_seen_at && (
+          <span className="text-muted-foreground"> · {lastSeen}</span>
+        )}
+        {!offline && wlState !== "idle" && (
+          <span className={wl.textClass}>
+            {" · "}
+            {tAgents(($) => $.workload[wlState])}
+            {counts ? (
+              <span className="font-mono tabular-nums text-muted-foreground">
+                {" "}
+                {counts}
+              </span>
+            ) : null}
+          </span>
+        )}
+      </span>
+    </ListGridCell>
+  );
+}
+
+// Per-row cost — only renders a 7d total + delta vs the prior 7d, so we
+// only need 14 days of usage. Previously this fetched a 180-day window to
+// share the cache key with the runtime-detail page, but that turned the
+// list page into N × 180d in-line aggregations against `task_usage` (one
+// per runtime row) and dominated DB load for this view. Detail still
+// fetches its own 180d window on navigation; the cold-load difference for
+// detail is one extra request, while the steady-state savings on the list
+// page are large.
+const COST_CELL_DAYS = 14;
+
+export function CostCell({ runtimeId }: { runtimeId: string }) {
+  const { t } = useT("runtimes");
+  const tz = useViewingTimezone();
+  const { data: usage = [] } = useQuery(
+    runtimeUsageOptions(runtimeId, COST_CELL_DAYS, tz),
+  );
+  const cost7d = useMemo(() => computeCostInWindow(usage, 7, tz), [usage, tz]);
+  const costPrev7d = useMemo(
+    () => computeCostInWindow(usage, 7, tz, 7),
+    [usage, tz],
+  );
+  const delta = pctChange(cost7d, costPrev7d);
+
+  if (usage.length === 0) {
+    return (
+      <div className="w-full text-right">
+        <span className="text-xs text-muted-foreground/50">—</span>
+      </div>
+    );
+  }
+  const fmt = cost7d >= 100 ? `$${cost7d.toFixed(0)}` : `$${cost7d.toFixed(2)}`;
+  const deltaTone =
+    delta == null
+      ? "text-muted-foreground"
+      : delta > 0
+        ? "text-warning"
+        : delta < 0
+          ? "text-success"
+          : "text-muted-foreground";
+  const deltaLabel =
+    delta == null
+      ? null
+      : delta === 0
+        ? t(($) => $.list.cost_delta_flat)
+        : `${delta > 0 ? "↑" : "↓"}${Math.abs(delta)}%`;
+  return (
+    <div className="flex w-full flex-col items-end leading-tight">
+      <span className="text-sm font-medium tabular-nums">{fmt}</span>
+      {deltaLabel && (
+        <span className={`text-[11px] tabular-nums ${deltaTone}`}>
+          {deltaLabel}
+        </span>
+      )}
+    </div>
+  );
+}
+
+export function CliCell({ runtime }: { runtime: AgentRuntime }) {
+  if (runtime.runtime_mode === "cloud") {
+    return <span className="text-xs text-muted-foreground/50">—</span>;
+  }
+  const meta = runtime.metadata as Record<string, unknown> | null;
+  // `version` is the agent's own underlying CLI tool version — distinct per
+  // provider (e.g. "2.1.5 (Claude Code)", "codex-cli 0.118.0", "0.42.0").
+  // The separate `cli_version` is the shared multica daemon CLI, identical
+  // for every runtime on one machine; surfacing it here made all agents
+  // show the same number (#3838). The daemon CLI version and its update
+  // prompt belong to the machine — they live in the machine meta strip and
+  // the detail page's UpdateSection, not on a per-agent row.
+  const version =
+    meta && typeof meta.version === "string" ? meta.version : null;
+
+  if (!version) {
+    return <span className="text-xs text-muted-foreground/50">—</span>;
+  }
+
+  return (
+    <div className="flex min-w-0 items-center text-xs">
+      <span className="truncate font-mono text-muted-foreground">
+        {version}
+      </span>
+    </div>
+  );
+}
+
+// Stacks up to 3 agent avatars, then a "+N" pill if more bind to this
+// runtime. Each avatar uses the wrapping ActorAvatar so hover automatically
+// surfaces AgentProfileCard.
+function AgentStack({ agentIds }: { agentIds: string[] }) {
+  if (agentIds.length === 0) {
+    return <span className="text-xs text-muted-foreground/50">—</span>;
+  }
+  const visible = agentIds.slice(0, 3);
+  const extra = agentIds.length - visible.length;
+  return (
+    <div className="flex items-center -space-x-1.5">
+      {visible.map((id) => (
+        <span
+          key={id}
+          className="inline-flex rounded-full ring-2 ring-background"
+        >
+          <ActorAvatar
+            actorType="agent"
+            actorId={id}
+            size={22}
+            enableHoverCard
+          />
+        </span>
+      ))}
+      {extra > 0 && (
+        <span className="inline-flex h-6 w-6 items-center justify-center rounded-full bg-muted text-xs font-medium text-muted-foreground ring-2 ring-background">
+          +{extra}
+        </span>
+      )}
+    </div>
+  );
+}
+
+export function RuntimeRowMenu({
+  runtime,
+  wsId,
+  canDelete,
+}: {
+  runtime: AgentRuntime;
+  wsId: string;
+  canDelete: boolean;
+}) {
+  const { t } = useT("runtimes");
+  const [deleteOpen, setDeleteOpen] = useState(false);
+  // Delete is currently the only row action; if the row can't run it, drop
+  // the kebab entirely so the column doesn't render an empty popover. The
+  // self-healing case (local + online) is the runtime-detail parity fix —
+  // see isSelfHealingRuntime for the rationale.
+  const selfHealing = isSelfHealingRuntime(runtime);
+
+  if (!canDelete || selfHealing) {
+    return <span aria-hidden />;
+  }
+
+  return (
+    <>
+      <DropdownMenu>
+        <DropdownMenuTrigger
+          render={
+            <Button
+              variant="ghost"
+              size="icon-sm"
+              aria-label={t(($) => $.list.row_actions_aria)}
+              onClick={(e) => e.stopPropagation()}
+              onKeyDown={(e) => e.stopPropagation()}
+            />
+          }
+        >
+          <MoreHorizontal className="h-4 w-4 text-muted-foreground" />
+        </DropdownMenuTrigger>
+        <DropdownMenuContent
+          align="end"
+          className="w-40"
+          onClick={(e) => e.stopPropagation()}
+        >
+          <DropdownMenuItem
+            variant="destructive"
+            onClick={() => setDeleteOpen(true)}
+            title={t(($) => $.list.delete_permission_hint)}
+          >
+            <Trash2 className="h-3.5 w-3.5" />
+            {t(($) => $.list.delete_action)}
+          </DropdownMenuItem>
+        </DropdownMenuContent>
+      </DropdownMenu>
+      <DeleteRuntimeDialog
+        open={deleteOpen}
+        onOpenChange={setDeleteOpen}
+        runtime={runtime}
+        wsId={wsId}
+        onDeleted={() => {
+          setDeleteOpen(false);
+          toast.success(t(($) => $.detail.toast_deleted));
+        }}
+      />
+    </>
+  );
+}
+
+// ---------------------------------------------------------------------------
+// List
+// ---------------------------------------------------------------------------
+
 export function RuntimeList({
   runtimes,
   updatableIds,
@@ -87,8 +462,7 @@ export function RuntimeList({
 
   const { t } = useT("runtimes");
   const wsId = useWorkspaceId();
-  const slug = useWorkspaceSlug();
-  const navigation = useNavigation();
+  const wsPaths = useWorkspacePaths();
   const user = useAuthStore((s) => s.user);
 
   const { data: agents = [] } = useQuery(agentListOptions(wsId));
@@ -135,36 +509,93 @@ export function RuntimeList({
     }));
   }, [runtimes, memberById, workloadIndex, isAdmin, user]);
 
-  const columns = useMemo(
-    () =>
-      createRuntimeColumns({
-        showOwner,
-        wsId,
-        now,
-        t,
-      }),
-    [showOwner, wsId, now, t],
-  );
-
-  const table = useReactTable({
-    data: rows,
-    columns,
-    getCoreRowModel: getCoreRowModel(),
-    enableColumnResizing: true,
-    // Pin the kebab column right so it stays accessible during horizontal
-    // scroll — matches the pattern in Linear / Notion / GitHub.
-    initialState: { columnPinning: { right: ["actions"] } },
-  });
-
   return (
-    <DataTable
-      table={table}
-      onRowClick={(row) => {
-        if (!slug) return;
-        navigation.push(
-          paths.workspace(slug).runtimeDetail(row.original.runtime.id),
-        );
-      }}
-    />
+    <div className="overflow-x-auto @container">
+      <ListGrid
+        className={`${GRID_COLS} @2xl:min-w-[var(--rtc-minw)]`}
+        style={columnTrackVars(showOwner)}
+      >
+        <ListGridHeader>
+          <ListGridHeaderCell>
+            {t(($) => $.list.col_runtime)}
+          </ListGridHeaderCell>
+          <ListGridHeaderCell>{t(($) => $.list.col_health)}</ListGridHeaderCell>
+          {showOwner ? (
+            <ListGridHeaderCell className="hidden @2xl:flex">
+              {t(($) => $.list.col_owner)}
+            </ListGridHeaderCell>
+          ) : (
+            <ListGridHeaderCell className="hidden px-0 @2xl:flex" />
+          )}
+          <ListGridHeaderCell className="hidden @2xl:flex">
+            {t(($) => $.list.col_agents)}
+          </ListGridHeaderCell>
+          <ListGridHeaderCell className="hidden @2xl:flex" align="right">
+            {t(($) => $.list.col_cost)}
+          </ListGridHeaderCell>
+          <ListGridHeaderCell className="hidden @2xl:flex">
+            {t(($) => $.list.col_cli)}
+          </ListGridHeaderCell>
+          <span aria-hidden="true" />
+        </ListGridHeader>
+        {rows.map((row) => (
+          <ListGridRow
+            key={row.runtime.id}
+            render={<AppLink href={wsPaths.runtimeDetail(row.runtime.id)} />}
+          >
+            <RuntimeNameCell runtime={row.runtime} />
+            <HealthCell
+              runtime={row.runtime}
+              workload={row.workload}
+              now={now}
+            />
+            {showOwner ? (
+              <ListGridCell className="hidden gap-1.5 @2xl:flex">
+                {row.ownerMember ? (
+                  <>
+                    <ActorAvatar
+                      actorType="member"
+                      actorId={row.ownerMember.user_id}
+                      size={18}
+                    />
+                    <span className="min-w-0 truncate text-xs text-muted-foreground">
+                      {row.ownerMember.name}
+                    </span>
+                  </>
+                ) : (
+                  <span className="text-xs text-muted-foreground/50">—</span>
+                )}
+              </ListGridCell>
+            ) : (
+              <ListGridCell className="hidden px-0 @2xl:flex" />
+            )}
+            <ListGridCell className="hidden @2xl:flex">
+              <AgentStack agentIds={row.workload.agentIds} />
+            </ListGridCell>
+            <ListGridCell className="hidden @2xl:flex">
+              <CostCell runtimeId={row.runtime.id} />
+            </ListGridCell>
+            <ListGridCell className="hidden @2xl:flex">
+              <CliCell runtime={row.runtime} />
+            </ListGridCell>
+            <ListGridCell className="justify-end px-0">
+              <span
+                onClick={(e) => {
+                  e.preventDefault();
+                  e.stopPropagation();
+                }}
+                className="flex items-center"
+              >
+                <RuntimeRowMenu
+                  runtime={row.runtime}
+                  wsId={wsId}
+                  canDelete={row.canDelete}
+                />
+              </span>
+            </ListGridCell>
+          </ListGridRow>
+        ))}
+      </ListGrid>
+    </div>
   );
 }
